+++ packages/plugin-html/web/src/getToolbarButtons.tsx
@@ -2,9 +2,18 @@ import React from 'react';
 import type { ToolbarButton } from 'ricos-types';
 import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
 import { EditHtmlButton } from './toolbar/EditHtmlButton';
+import { DimensionSliderButton } from './toolbar/DimensionSliderButton';
 import { TIPTAP_HTML_TYPE } from 'ricos-content';
+import { MAX_HEIGHT_INPUT, MIN_WIDTH, MAX_WIDTH, MIN_HEIGHT, MAX_HEIGHT } from './defaults';
+import { HeightIcon, WidthIcon } from 'wix-rich-content-plugin-commons';
 
 export const getToolbarButtons = (config): ToolbarButton[] => {
+  const {
+    minHeight = MIN_HEIGHT,
+    maxHeight = MAX_HEIGHT,
+    maxWidth = MAX_WIDTH,
+    minWidth = MIN_WIDTH,
+  } = config || {};
   return [
     {
       id: 'editHtml',
@@ -14,6 +23,46 @@ export const getToolbarButtons = (config): ToolbarButton[] => {
       renderer: toolbarItem => <EditHtmlButton toolbarItem={toolbarItem} />,
       attributes: {
         nodeAttrsInSelection: nodeAttrsInSelectionResolver,
+      },
+    },
+    {
+      id: 'htmlWidth',
+      command: ({ data, editorCommands }) => {
+        editorCommands.chain().updateAttributes(TIPTAP_HTML_TYPE, data).run();
+      },
+      icon: WidthIcon,
+      tooltip: 'ChangeDimensions_Width_Tooltip',
+      renderer: toolbarItem => (
+        <DimensionSliderButton
+          toolbarItem={toolbarItem}
+          dimension={'width'}
+          min={minWidth}
+          max={maxWidth}
+          inputMax={MAX_HEIGHT_INPUT}
+        />
+      ),
+      attributes: {
+        selectedNode: selectedNodeResolver,
+      },
+    },
+    {
+      id: 'htmlHeight',
+      command: ({ data, editorCommands }) => {
+        editorCommands.chain().updateAttributes(TIPTAP_HTML_TYPE, data).run();
+      },
+      icon: HeightIcon,
+      tooltip: 'ChangeDimensions_Height_Tooltip',
+      renderer: toolbarItem => (
+        <DimensionSliderButton
+          toolbarItem={toolbarItem}
+          dimension={'height'}
+          min={minHeight}
+          max={maxHeight}
+          inputMax={MAX_HEIGHT_INPUT}
+        />
+      ),
+      attributes: {
+        selectedNode: selectedNodeResolver,
       },
     },
     {
@@ -32,5 +81,16 @@ const nodeAttrsInSelectionResolver = {
       return content[0].attrs;
     }
     return false;
+  },
+};
+
+const selectedNodeResolver = {
+  id: 'selectedNode',
+  resolve: content => {
+    if (Array.isArray(content) && content.length > 0) {
+      return content[0];
+    } else {
+      return undefined;
+    }
   },
 };