+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 
 - `ricos-events`
   - [#4149](https://github.com/wix/ricos/pull/4149) getAllTopics; publishOnce; duplication errors
+- `html-plugin`
+  - [#4153](https://github.com/wix/ricos/pull/4153) tiptap html width and height sliders
 
 ## 8.71.38 (Jul 9, 2022)